ignore generated doc tags
A previous PR dealt with this (febd9162) but the generated files show up
for me as doc/tags, not docs/tags. If it was working for the author of
that PR I don't want to break it for him and anyone that was working
for, so I'll keep the existing rule as well.